app/requirements/development.txt
alembic==1.7.4
aniso8601==9.0.1
apispec==5.2.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachelib==0.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.0.2
Flask-Admin==1.5.8
flask-apispec==0.11.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-JWT-Extended==4.4.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
imagekitio==2.2.8
importlib-metadata==4.11.3
importlib-resources==5.6.0
inflection==0.5.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.2
jsonschema==4.4.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.15.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
speaklater==1.3
SQLAlchemy==1.4.26
stack-data==0.2.0
terminado==0.13.3
tinycss2==1.1.1
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
webargs==8.1.0
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
WTForms==2.3.3
zipp==3.8.0